TYP: move ``flatiter`` stubs from ``__init__.pyi`` to ``_core/multiarray.pyi``
